Fixed Liked Songs bug
I fixed a bug where liked songs needed to be clicked twice to update the searchResults. To fix this I moved the displayLikedSongs call in NavBottom into a useEffect, with depndencies of likedSongs and displayLikedSongs. This introduced another bug where triggering that useEffect caused the component to re-render, which then caused the useEffect to trigger once again, and so on and so forth. To fix this I wrapped displayLikedSongs in a useCallback in Dashboard.js.