Custom hook useLikedSongs now works.
It has params spotifyApi, convertDuration, and userID. userID is not used in the api call, but merely as a buffer so that the api isn't called when useAuth is not yet done retrieving the accessToken, which is needed for the call. This is needed to avoid an Error 401: no token provided. I may need to find a less 'hacky' way to do this but for now this works.