Updated useLikedSongs to use useContext
useLikedSongs now uses useContext to access the spotifyApi and convertDuration function Dashboard.js. This is better than passing down the two as parameters because most if not all of my custom hooks in the future will need spotifyApi, and passing it down every time I use a custom hook would make the code look repetitive and cluttered.