Updated useLikedSongs useEffect dependency
Similar to the infinite loop with the previous iteration of userUserPlaylists, useLikedSongs API was being called infinitely. To fix this I created a simple 0 or 1 state variable in Dashboard that changes only when the user likes or unlikes a song. I am now passing that down as an argument to useLikedSongs instead of passing down changeTrackLikeStatus.