Clicking on Search now focuses on the searchBar
I did this by using useref and assigning it to the searchBar input and then passing down a props function to Nav that focuses on the searchBar on click. On a side note: usePlayingUris is incomplete. Clicking on a track works but clicking on PLay within Your Library does not work. Need to fix