Fixed the bug where the currentUris state in Dashboard was one render behind
To fix this, I moved the logic that I used in usePlaylistTracks custom hook and put it directly inside Dashboard. There were too many variables being passed down to nested hooks, and that caused unneccessary code complexity and unreadability. Thus, I chose to forgo using the usePlaylistTracks custom hook. However, there may be a use for the hook in the future so I will keep it on hand.