Fixed longstanding pause bug
When the player is paused and I try to play another song from the same list of songs, the new song will now play. I fixed this by combining all the states in Player.js to a useReducer instead. This allowed the SpotifyPlayer to access all the values it needed from one single source(the reducer state) rather than me having to manually update and pass down multiple states.