Moved useReducer logic in Player into a global context
The context is accessible to all components as its wrapped around App.js. Currently Dashboard, Track, and Player make use of its values uris[], offset(int), and play(bool) to conditionally display/play the correct tracks. However, there are bugs that need to be fixed. Pausing and trying to play a song won't play the song until another one is selected. On startup, the player play button must be pressed directly (clicking on a track play button) does not start the player.